Add ast equality assertion helper
Add a testing helper function for determining if two `ast.AST` nodes are
equal by value.

Using the `==` operate does not work with objects of type `ast.AST`
because it compares if they are the same object in memory. It would be
nice if `ast.AST` had a proper `__equal__` method but it doesn't. Maybe
I ought to submit a patch to Python?