Remove test helpers
Remove the test helpers module.

Story time!

Originally I was going to use the ast module for Arborista, and I got as
far as implementing a working transformer and a transformation which
removed statements after a return, but then I realized there were some
problems...

The first problem was that the ast module didn't have a good native way
of converting the ast back to a string.

"This isn't that bad", I thought. I can just write some dumps (dump
string) methods for nodes. It would be tedious but doable.

But then I ran Arborista on some code with a comment in it and I had a
realization that there was a worse problem: ast doesn't preserve
comments. I already knew it didn't preserve whitespace but I didn't
care much about that because I figured people would just run a
formatter afterwards, but not preserving comments was not acceptable.

I saw two options at the time:

1. Hackily parse the code and make a comment node myself.
2. Make a parser and lexer for Python or use a tool to generate one.

I didn't like either of these options all that much so I stalled and
didn't work on Arborista for a long time (7 months).

I can't remember if I looked around for other parsers at the time or
not but eventually I stumbled upon LibCST by Instagram. One of the
declared use cases for LibCST is "codemod" or automated refactoring.

So, it turns out others have already worked on this problem. That's
okay, they didn't gain much traction in the CS community so this is
still worth doing for that reason. And I think they set their sights
too low...

I didn't really want to use LibCST completely because it didn't look
like it was under development anymore and depending on it seemed
fragile.

A solution to this dilemma was introduced to me in "Clean Code", a book
by Uncle Bob. (Thanks Bob!) The solution is to wrap all usage of third
party code. This has several advantages:
1. You can replace usage much easier since the third party code is all
in one place.
2. You isolate the dependency on the package.
3. Your code which uses the wrapper is simpler and easier to read. For
example, it gets to call methods with names which make sense in your
context.

So, the plan now is to use LibCST and that is the story of why I am
deleting this file.