Update README.md with another wild idea
Update the README.md to include an idea about generating lexers,
parsers, and nodes from BNF or EBNF specifications of languages.

As a side note, while I see Arborista as have potential for
text languages, it is important to think about other kinds of tree
(or maybe even graph structures) that we can apply these sorts of
methods to. The universe is hypergraph rewrites according to
Stephan Wolfram...