Add libcst dependency
Add a dependency on libcst for parsing Python concrete syntax trees. I
would like to remove this dependency eventually but it convenient to
lean on libcst for now.

We might want to figure out a lower bound on the version of libcst
needed at some point and maybe set up an automated test for it?