Simplify Python block init
Simplify the initalization of Python blocks to just take a single
statements arguement. It used to take two arguemnts, the first statement
and the rest of the statements. The reasoning for that was so that it
wouldn't be possible to make a block with no statments. The second
argument was optional, but this now poses a problem. The problem is that
block needs an indent argument too. The indent argument should not be
optional but it also doesn't make much sense for it to come before the
statements arguemnts. So, for now it seems better to loosen the
initialization.