Add indent to Python block
Blocks are indented by a string. Unfortunately the off-sidedness of
Python (the syntactic significance of indentation) means that conversion
from a CST to a string will be context sensitive. It might make sense to
do all context sensitive processing inside of the lexer and maintain the
parser as context-free so that we can generate it with a parser
generator from EBNF or better yet named-ENBF (NEBNF?).