Adjust main to parse, transform, and then deparse
Adjust main to parse the file argument into as string then parse the
string into a module. The module is then transformed. The transformed
module is deparsed which involves turning it into a file and then
writing the file.

These three steps should really be three suites of transformations which
should be performed in succession. Id est, the transformer should have
the ability to run different sets of transformations with different
policies. The parsing should be a set of transformations which are
applied once, then the transformations which improve the Python nodes
should but applied in a repeated fashion until there are no more
improvements to be made, and then lasty the deparsing transformations
should be performed once.

It might be worth it to distinguish between transformations which have
side effects and those that don't in order to do lazy evaluation.