Add trim after return transformation
Add a transformation for trimming small statements after a return
statement in a simple statement.

It also makes sense to have a similar transformation for trimming
statements after a return statement in a block. But if these two
transformations are not made to work well together, then it could lead
to inefficency.

There are two cases for handling trimming in blocks and simple
statements. The first case is that they are seperate transformations.
In this case, the lack of informations sharing means that you might be
checking for the same condition multiple times. The second case is that
one transfomation handles both types of nodes. Again it seems like
without added coordination, there would be duplication of condition
checking.

I think this comes down to a problem of reference frame. Currently,
transformations must be framed from the perspective of the parent node.
This sort of transformation is actually best framed from the
perspective of the child node (the return statement). Coordinating
transformations which are not top down might be more difficult, but
seems like it might be a direction that Arborista should go in.