Change trim after return reference frame
Change the reference frame for the trim after return transformation from
the parent node to the return statement. It is easier to describe the
transformation from the perspective of the return statement than from
the perspective of the parent statement and it is easier to propogate
the information up rather than searching down arbitrary levels.

This still only handles simple statements and only one level of nesting.

It should be generalized to climb multiple levels in the hierarchy and
also to be any flow interruption statement (e.g. break statements too).
In addtion other types of structures such as block statements should be
supported.

The plan right now is to only have one transformation run at a time, but
it makes sense to parallelize transformation running in which case
locking of nodes, deadlock detection, and intelligent deadlock handling
need to be added too.