Add node children iteration
Add iteration of children nodes for different node types. The need for
child iteration comes from the transformer needing to be able to walk
the tree.

Some nodes probably need have their data converted into new node types
so that the data can be transformed. For example, the contents of the
file node should probably be converted into a node so that it can parsed
from a string into a module by a parsing transformation. All this is to
say that more work needs to be done for adding child iteration.