Add testing helpers package
Add a package for helper code for testing. It is not clear where this
type of code should live and the iternet has little advise to offer on
the subject. For now, lets try this out and we can reasses later.

The other open question is should there be tests for testing helpers?