Add testing helper for checking a set of nodes
Add a testing helper for checking that a set of nodes matches an
expected set. Right now nodes are mutable and so they cannot be hashable
which means that they cannot be stored in a set. We should do something
about this so we don't have to have such an ugly helper method.