Change file contents to be a node
Change the type of the file contents to be a node.

This is needed in order to be able to convert parsing and deparsing into
transformations. (This means that some of the existing parsing and
deparsing code which was changed here in maybe not the most elegant way
will be removed.)