Add node equality operator
Add the equality operator for nodes and make node an abstract base
class. The reason for doing this is to ensure that all nodes have an
appropriate equality operator defined. (It would probably be better
however to define a generic equality operator which checks that all
data is the type and data for the two objects being compared.)