Use equal attributes decorator
Use the equal attributes decorator from Seligimus. There are a lot
places where this can't be used yet the decorator does not have the
ability to exclude certain attributes. For example, with Python nodes
we do not want to include the parent node when checking equality. (It
is also possible that this is the wrong behavior and that we _do_
actually want to include the parent node...)